+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Auto label points on volcano plot.
 - Gsea plot now loads preranked data.
 - Fixed p-value and threshold filter on volcano by adding proper bounds and removing unneccessary log conversion.
 - Bug fixes and updates.